vyos-kernel: disable kernel debugging for all configs
Disable kernel debuging in the kernel config file for 586-vyos and
586-vyos-virt kernels.  Removing debugging support adds less kernel
based overheads for something that will be rarely used in a production
environment. This was disabled for the amd64 flavour in commit
1cc1b7cd4abd0f9cb6a51fa2f266ed32e685e3da

Possibly need to look at introducing a debug build for non production
systems.

Bug #283 http://bugzilla.vyos.net/show_bug.cgi?id=283